Added double cast in all models to trigger null pointer catch for default.  Added code to accomodate for user setting up initial configurations and changing state colors

+++ src/Models/Fire.java
@@ -37,7 +37,7 @@ public class Fire extends BaseModel {
 		List<Integer> myInts = new ArrayList<>(Arrays.asList(EMPTY, TREE, BURNING));
 		initializeMaps(myStates, myInts, myColors);
 		try {
-			getParameterValuesMap().put("probCatch", parameters.get("probCatch"));
+			getParameterValuesMap().put("probCatch", (double) parameters.get("probCatch"));
 		} catch (NullPointerException e) {
 			getParameterValuesMap().put("probCatch", (MIN_PROB_CATCH + MAX_PROB_CATCH) / 2);
 		}
